requirements.txt
# -*- coding: utf-8 -*-
# 工程预装模块
# 需要额外的python包，可直接在文件后面添加
# 请确保指定的包和版本号，可通过pip安装

# blueapps requirement

redis==3.5.3
Django==3.2.4
celery==4.4.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0

# jwt
pyCryptodome==3.9.7
PyJWT==1.7.1
#cryptography==2.7
mysqlclient==2.0.3
MarkupSafe==1.1.1
Mako==1.1.4
requests==2.25.1
python-json-logger==2.0.1
whitenoise==5.2.0
gunicorn==19.6.0

blueapps==4.1.1

certifi==2021.5.30; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
chardet==4.0.0; python_version >= "2.7" and python_full_version < "3.0.0" and python_version < "3" or python_full_version >= "3.6.0" and python_version < "3"
charset-normalizer==2.0.2; python_full_version >= "3.6.0" and python_version >= "3"
curlify==2.2.1
idna==2.10; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
pytz==2021.1
typing-extensions==4.1.1
typing==3.10.0.0; python_version >= "2.7" and python_full_version < "3.0.0" and python_version < "3.5" or python_full_version >= "3.4.0" and python_version < "3.5" or python_version == "2.7"
urllib3==1.26.6; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"

# 以下为自定义模块
kubernetes
pymongo
djangorestframework==3.14.0
django-redis-cache
hiredis
importlib-metadata==4.8.3
SQLAlchemy==1.4.50
